fixed pr issues
